CRAS: audio thread - Fix race with stream add/remove/add.
If a client does stream add/remove/add very quickly, as Chrome 26 now
does, there is a window where the iodev can be re-used for the re-add
after being un-configured due to remove.  To resolve this make sure
the device is closed on errors, and set the iodev format from the
audio thread before opening it.

BUG=chrome-os-partner:17156
TEST=play mp4 video from SSD in Chrome.

Change-Id: Iaf37a147bcacd7e9476743850405b67f445d665c
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/41029