CRAS: audio thread - handle dev open errors better.
If a device fails to open, remove it from the device list and try
again.  Make sure that the audio thread doesn't run unless both
devices are open, check before starting the unified_io function.
add_thread can now return errors to the main thread so that it can
remove the iodev if there is an error.

BUG=chromium-os:37798
TEST=updated rclient unittest

Change-Id: I056155989e2c6152d8226ab84d0dae678ab6dcae
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/41170