CRAS: audio thread - Use device_active from unified_io.
The function device_active was recently added to check if there is an
active device on the thread.  Use that to re-work commit
361d9ae06db8421524aa9d279b244e10ca259564, previously this was achieved
by setting the device pointer to null when it is closed, now we can
check if the device is open.  Not setting the pointer to NULL allows
the thread to be re-used when another stream is added.

BUG=chrome-os-partner:17214
TEST=fast reloading of a page that plays audio or of an mp4 file in the media player.

Change-Id: I6fc4c5599efaa4e93494f18edd8c300696a5b02f
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/41219