CRAS: audio thread - Move buffer fill for unified audio.
Instead of padding the output buffer for unified audio when the stream
is created, do it whenever there are zero frames queued before
playback.  This fixes the case where the output buffer is exhausted
due to underrun.  Previously this would lead to a constant string of
underruns since the output buffer would never get re-padded.

BUG=chromium-os:35272
TEST=play audio through cras on z620, cause underrun with debug
version of chrome.  This fixes issues where unified audio would
constantly under-run and sound choppy on ChromeOS devices.

Change-Id: If295b379b2a7a98ec202996000191ee19531150b
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/41462
Reviewed-by: Chih-Chung Chang <chihchung@chromium.org>