CRAS: rstream - Limit number of requested frames.
Always ask the client for min_cb_level frames.  This gives the client
a more consistent callack level.  Chrome has a DCHECK to make sure
that this happens, this change will allow debug builds of chrome to
run unmodified.  Chrome would not ever provide more than this many
samples so there is no run-time change to the system because of the
request size changing.

BUG=chromium:165288
TEST=updagte rstream_unittest

Change-Id: I0ffbf5445ad4dc5c1fc6cb984933f77571501cda
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/42451