CRAS: Allow unified path to be used for play/capture.
Allow playback and capture streams to be setup with unified
parameters.  Playback streams are indicated by a null capture buffer
and capture streams by a null playback buffer.  This will allow for
chrome to use the same path for all stream types and reduce the amount
of duplicate code in its CRAS audio back end.  This means changes to
audio_thread to allow for a stream to drop an input or output while
running.

Use the new path from pcm_cras alsa plugin.

BUG=chromium-os:35272
TEST=play/record audio with test client, and with chrome on Link.

Change-Id: I71f01004b18013bdf77c2a0a18321fcdefd8416b
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/43253
Reviewed-by: Chih-Chung Chang <chihchung@chromium.org>