CRAS: Always create a "Speaker" node for the first internal alsa device.
Some machines (like lumpy) don't have a mixer control for the built-in
Speaker, so we would only create a node with the name "(default)". In
the future we want to let the user to select the output node, so we need
to give the node the correct name "Speaker".

BUG=chromium:160311
TEST=check default node name using cras_test_client --dump_server_info on lumpy
Change-Id: If0974bb472c08a0e99b84760e572d58ddd5cb332
Reviewed-on: https://gerrit.chromium.org/gerrit/43956
Commit-Queue: Chih-Chung Chang <chihchung@chromium.org>
Reviewed-by: Chih-Chung Chang <chihchung@chromium.org>
Tested-by: Chih-Chung Chang <chihchung@chromium.org>