CRAS: Use ionode for output selection instead of iodev.
The plugging state and priority in iodev are moved to ionode, and the
output selection logic is based on ionode instead of iodev.

In addition to the speaker node, we also always create an internal mic
node, so we always have an active input node.

BUG=chromium:160311
TEST=make check; check output of cras_test_client --dump_server_info on lumpy
Change-Id: I540f2a656dcc15b03e687a645b2682a6e9d31176
Reviewed-on: https://gerrit.chromium.org/gerrit/43958
Commit-Queue: Chih-Chung Chang <chihchung@chromium.org>
Reviewed-by: Chih-Chung Chang <chihchung@chromium.org>
Tested-by: Chih-Chung Chang <chihchung@chromium.org>