Move code that are common for all iodevs from alsa_io to iodev.
This moves the handling of plugging and selection from alsa specific
code to the general iodev code. This makes it unnecessary to duplicate
the same logic for each kind of iodev. It also makes it easier to
add more hooks when plugging and selection happens.

BUG=none
TEST=make check
Change-Id: I45bcb0d24d98c16cfc2ff09abee7e4b393532356
Reviewed-on: https://gerrit.chromium.org/gerrit/44294
Commit-Queue: Chih-Chung Chang <chihchung@chromium.org>
Reviewed-by: Chih-Chung Chang <chihchung@chromium.org>
Tested-by: Chih-Chung Chang <chihchung@chromium.org>