CRAS: Check that there is a real output device.
Check if there is an output device in the system.  Do this 5 seconds
after the server is started.  The five second delay is to allow for
udev enumeration to complete.  This log will be converted to an UMA
stat so that codec failures can be monitored.

BUG=none
TEST=start cras check that no message is present, remove codec driver
module from kernel, restart cras and see message in /var/log/messages.

Change-Id: I92d30159e01446248c85adb3ce7d6e038c134b85
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/44743
Reviewed-by: Chih-Chung Chang <chihchung@chromium.org>