CRAS: Add cras_alert facility.
This is a similar to the state_callback_list used in
cras_system_state. The differences are:

(1) We do not call the callbacks immediately. We mark an alert as
"pending" when we know the callbacks should be called, but only actually
do so at the end of the event loop. This collapses multiple callbacks into
one (for example, if there are multiple nodes added at the same time,
we will only fire "nodes changed" signal once)

(2) We added a "prepare" function which will be called before calling the
callbacks. This gives each alert a chance to update the system to a
consistent state before signalling the clients.

We will move cras_system_state to use cras_alert in the next change.

BUG=chromium-os:33666
TEST=make check
Change-Id: I2c98430754f7395dfa96f82cde82dfb5fac65f00
Reviewed-on: https://gerrit.chromium.org/gerrit/45188
Reviewed-by: Dylan Reid <dgreid@chromium.org>
Commit-Queue: Chih-Chung Chang <chihchung@chromium.org>
Tested-by: Chih-Chung Chang <chihchung@chromium.org>