CRAS: Allow iodev to specify a minimum buffer level.
This additional buffering can be used for devices that transfer data
in chunks > 1.  HDA or i2s devices allow for continuous transfer of
samples, USB devices grab samples in URB sized chunks.  If the buffer
level isn't filled with more than the chunk size when a URB needs to
be sent, the device will under-run.  This new buffer level can be used
to set the floor for the number of samples queued.  A subsequent
change will set the floor to a non-zero value for ALSA USB devices.

BUG=chrome-os-partner:17660
TEST=set min buffer level and play to USB device with a buffer size of
480 frames, no glitches.

Change-Id: I3dc3b88983ef50850d9089f7e16343e74f804d56
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/46320