CRAS: Set floor for USB audio buffering.
Don't let the audio buffer drop too low for USB devices.  This will
prevent a situation where the device needs a large number of samples
and under-runs because the buffer hasn't been filled yet.  Even if the
buffer is filled before the samples would be played according to the
audio clock, an underrun can occur if a large chunk of audio is
requested.  Often URB size is ~400 frames.  Pad an extra 768 to be
safe.

BUG=chrome-os-partner:17660
TEST=play audio with buffer size 480 to USB device before and after
change.  Before the change, it would glitch, after it plays smoothly.

Change-Id: I559f488c496cded2f17b865290372e374e1d6d51
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/46321