CRAS: a2dp - Prevent leak of supported formats.
The list of supported sample rates and number of channels is set when
setting the format.  This can happen multiple times during the life of
an iodev.  Free these structures before allocating them when setting
the format.  Don't set them at init time as the format must be set
before the device is opened.

BUG=none
TEST=none

Change-Id: I8f354cd1ca33b619e8e55b2b1c0f8921c9d0728c
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/48792
Reviewed-by: Hsinyu Chao <hychao@chromium.org>