CRAS: Remove automatic device selection.
Don't auto select devices when they are added.  Chrome can now handle
this.  Keep the first added device as default, it will be the
emtpy_iodev and is used to store streams when an active device is
removed.  Chrome will react to the removal and move the stream to a
new active device.

CQ-DEPEND=CL:49921
BUG=175798
TEST=plug unplug devices, make sure audio goes to the right place.

Signed-off-by: Dylan Reid <dgreid@chromium.org>
Change-Id: I937fd89141e8d404d173ccf2f585f2ae656b77c2
Reviewed-on: https://gerrit.chromium.org/gerrit/49812