CRAS: shm - Add parameter to get_writeable_frames
Allow the total number of frames to be specified in calls to
cras_shm_get_writeable_frames.  A following commit will use this to
limit the amount unified streams read to the amount that should be
passed to the stream instead of greedily taking everything that will
fit in the buffer.

BUG=chromium:238971
TEST=make check, no functional change yet.

Change-Id: Ib6753ab484c0ef4bb6f6015bb1bb4d71b15d30f8
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/50948