CRAS: audio_thread - rework wake up levels for capture.
Change the way the next wake up time is calculated for capture and
unified streams.  Instead of relying on the hardware device's level,
use the level of the stream that will be ready next.  This allows for
multiple capture streams with non-integer multiple buffer sizes to run
concurrently.

possibly_read_audio is changed to fill in a parameter with the number
of frames until the next stream is ready.  It now limits the number of
frames read from hardware to the amount that will fit in the stream
with the least buffer until its callback threshold is met.

unified_io is changed to not ignore captured samples if they don't
meet the callback threshold of the device.  Instead always process
samples as a stream using that device could have its level met by this
wakeup.

This leaves output alone, it can be changes similarly, but there isn't
any bugs provoking the change so it will be left for a later date.

BUG=chromium:238971
TEST=web audio demos and updated audio_thread unit tests.

Change-Id: I36735b9cdcb2b0dc0312f911b287ca126d531178
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/50950