CRAS: bluetooth - handle changes in transport state
When we receive a suspend from an A2DP device, release the transport
stream and drop the iodev; and when we receive a start from an A2DP
device, re-create the iodev and, if music is playing, re-acquire the
transport stream.

Rename the endpoint methods since they are called in multiple scenarios
now, and track the endpoint/transport relationship.

BUG=chromium:238740
TEST=power off Monster Clarity HD while playing music,
  power back on and reconnect,
  verify music transitions to internal speaker when device issues a suspend,
  and back to device when it issues the start.

Change-Id: Ie278f5addd5a5d3997b320e2c8421b6dc0c6c092
Signed-off-by: Scott James Remnant <keybuk@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/51070
Reviewed-by: Dylan Reid <dgreid@chromium.org>
Reviewed-by: Hsinyu Chao <hychao@chromium.org>