CRAS: audio_thread - remove all streams from audio thread.
Don't iterate the streams list from the main thread.  Instead send a
message to the audio thread to remove all the streams.  The audio
thread can modify this list, it isn't safe to access it from the main
thread.

Streams should only be added to a running thread, check that a thread
is running before adding adding a stream and only remove streams from
a running thread.

BUG=none
TEST=switch devices while playing multiple streams, make sure all
migrate.

Change-Id: I148304813e72bcab3af76bfbaca29bdd372a5107
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/55592
Reviewed-by: Chih-Chung Chang <chihchung@chromium.org>