CRAS: audio_thread - Add loopback device support.
Push samples to the loopback device after mixing, but before applying
the DSP.  When a loopback stream is added, the loopback device will be
"opened" and each time data would be rendered, it is also passed to
the loopback device.

BUG=chromium:247260
TEST=updated audio_thread_unittest, play and record audio.  More tests
will come as the remaining parts of loopback are added to the system.

Change-Id: Ib00fa32b1aa8cae939b61661ba62b52cc4c40b91
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/60239
Reviewed-by: Chih-Chung Chang <chihchung@chromium.org>