CRAS: audio_thread - Mix all available samples.
After polling playback clients for more output samples, check how many
samples we have (ignore empty clients) and add that much data to the
output device.  Before this change, if a client wrote any data, the
write_limit would be mixed, even if write_limit was greater than the
amount of audio in the client's buffer.

BUG=chromium:247160
TEST=updated audio_thread_unittest

Change-Id: I42375110015eed503b0199d33f730503fae71f51
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/60243
Reviewed-by: Chih-Chung Chang <chihchung@chromium.org>