CRAS: audio_thread - Support loopback without output.
Add the ability to have a loopback stream running when no output
streams are.  Before the loopback stream would stall until a new
output started.  Instead fill it with zeros in accordance with the
stream's timing paramenters.  This also limit then number of
functional loopback streams to one, but that is all that is required
currently.

BUG=chromium:247160
TEST=use test client to record loopback stream without any output and
check that it is filled with zeros.

Change-Id: Ia7a8ec71b1163f3b5f182bf392450da3d3ed7d70
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/60244
Reviewed-by: Chih-Chung Chang <chihchung@chromium.org>