CRAS: dbus - Re-arm timeout timer if it fires.
DBUS expects the timer to fire repeatedly until it is removed.  If the
timer isn't rearmed a double-free might result as it is cancelled
after it expires and will be cancelled again when dbus calls remove.
Also re-work the add, remove, and togle functions.  Add clears the
timer if it already exists, remove checks that the timer exists before
attempting to remove it, and toggle checks if the timer is enabled
before taking an action.

BUG=none
TEST=build and run cras on Link and Spring.

Change-Id: I71ebca6f7c3723050374ee7cd35b89874c1f2625
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/60710
Reviewed-by: Chih-Chung Chang <chihchung@chromium.org>