Add dependency between installation rules.
Because of limitations in automake, there is a race between
installing libcras and libtool's relinking of the modules. This
occasionally causes the Chrome OS builders to fail, and need to
retry this package.

This change forces a dependency between the installation steps
to prevent the race condition. Tested by adding a sleep in the
generated install-lib_LTLIBRARIES rules, to produce the race every
time.

BUG=chromium:257634
TEST=Manual

Change-Id: I39a51d85fbbbe7e8b9a1f48441a32813a4e5c5ba
Reviewed-on: https://gerrit.chromium.org/gerrit/61026
Reviewed-by: Dylan Reid <dgreid@chromium.org>
Tested-by: Chris Wolfe <cwolfe@chromium.org>
Commit-Queue: Chris Wolfe <cwolfe@chromium.org>