CRAS: empty iodev - Report that cb_thresh samples are queued.
Since the input and output paths have been joined in the unified_io
function, frames_queued can always return the callback threshold for a
frame count.  If used as input this is no change, when used as output,
this will indicated that more samples should be read from the attached
streams.

BUG=chromium:247160
TEST=select empty_iodev and play output to it.

Change-Id: I086e7bf5e278e1dc5affffd3c08a0cd59a84303f
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/61200
Reviewed-by: Chih-Chung Chang <chihchung@chromium.org>
Reviewed-by: Hsinyu Chao <hychao@chromium.org>