Replace asound.state with UCM init for link.
This replaces the asound.state file with a set of configuration
commands sent by UCM.  This method is less fragile.  The asound.state
file fails whenever a control index changes.  The UCM method will only
fail if a control that is being set is removed.

BUG=258646
TEST=boot link and check controls values.  Repeatedly load this config
with the print for DSP queue overflow enabled and make sure that no
overflow happens.

Change-Id: I9614ff4d6c72f1c0ede048e7c1d7b41336349a6a
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/61760
Reviewed-by: Chih-Chung Chang <chihchung@chromium.org>