Replace asound.state with UCM init for mario.
This replaces the asound.state file with a set of configuration
commands sent by UCM.  This method is less fragile.  The asound.state
file fails whenever a control index changes.  The UCM method will only
fail if a control that is being set is removed.  This removes setting
some controls that no longer exist.

BUG=258646
TEST=boot mario and check controls values.

Change-Id: I5a4da02e9f94c3ea19a3a863ad887ee9f23dec22
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/61757
Reviewed-by: Chih-Chung Chang <chihchung@chromium.org>