Replace asound.state with UCM init for spring.
This replaces the asound.state file with a set of configuration
commands sent by UCM.  This method is less fragile.  The asound.state
file fails whenever a control index changes.  The UCM method will only
fail if a control that is being set is removed.

Also disable DAC2 in the mixers it isn't used.

BUG=258646
TEST=boot spring and check controls values.

Change-Id: I7bdab0b858b33d29c2e880bd31cfa2c6b3d454c4
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/61759
Reviewed-by: Chih-Chung Chang <chihchung@chromium.org>