CRAS: alsa_io - Avoid naming HDMI as internal speaker.
On falco, the HDMI output is the only device on a separate ALSA card.
When creating a default node for that card, avoid naming it as a
built-in speaker and instead make it an HDMI output.  This gives it
the correct priority and prevents it from being the default output if
it wins the race at startup.

BUG=chrome-os-partner:20980
TEST=list outputs on falco, check there isn't two "internal speaker"
outputs.

Change-Id: If1e7cd4cc95e4c17667f3459a8ed27d00e3de166
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/62349
Reviewed-by: Chih-Chung Chang <chihchung@chromium.org>