CRAS: audio_thread - only pad output buffer for unified streams.
Just because an input device is open doesn't mean that there is a need
to pad the output buffer.  That is only needed for unified streams.
Now that there is only one audio thread in the system and unrelated
input and output devices can be attached to it, padding will cause
unneeded glitches.

BUG=chromium:263585
TEST=manual, play/record/unified on Falco and Link.

Change-Id: I414c18f104ee34c25034b7de341031639ca92cc2
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/63390