CRAS: audio_thread - Remove volume/capture gain init from open_dev.
The ALSA open_dev calls init_device_settings which either sets volume
and volume limits (for output devices), or capture gain and capture
gain limits (for input devices).

Because the volume and capture gain are controls specific to each
device node, we wish to separate them from any audio processing that
occurs on the audio thread (including open_dev). This is necessary if
we wish to later on decouple the audio thread from the device
list/server.

First we remove init_device_settings from open_dev. This will mean
that the volume/capture gain is not reset correctly if a stream is
opened on a previously closed device (this is when open_dev is
normally called).

However, by moving the volume/capture gain calls to cras_rclient
whenever the client connects to a new stream, the volume/capture gain
will still be reset correctly in the above scenario.

Therefore in practice, the user experience should be consistent with
before.

The ALSA unit test has been updated to reflect this change. The
rclient unit test has been updated because of segmentation faults
(bogus calls to set_volume and set_capture_gain).

BUG=chromium:264890
TEST=Ran gmerge adhd and tested opening streams, plugging in
headphones and changing volume. Top-level behaviour is still
consistent.

Change-Id: I9f22b1f55c8fb59bfa516af3d6b8bfb6fa4b8177
Reviewed-on: https://gerrit.chromium.org/gerrit/64675
Reviewed-by: Dylan Reid <dgreid@chromium.org>
Commit-Queue: Adrian Li <alhli@chromium.org>
Reviewed-by: Adrian Li <alhli@chromium.org>
Tested-by: Adrian Li <alhli@chromium.org>