CRAS flags: Add AutoUnplugInputNode flag
When AutoUnplugInputNode flag is set in UCM, at the time a mic
jack is plugged/unplgged, internal mic will be set to
unplgged/plugged.

This is a work around for the audio device selection UI
to have reasonable behavior on platforms that we cannot
suppress auto-mic.

What user will observe is that the input device is auto-selected
and only the active one is displayed on UI.

BUG=chromium:267355
TEST=Manual set 'AutoUnplugInputNode' flag to true on ZGB,
plug and unplug mic jack.

Change-Id: Ia2402ce26e5a2673511cc651b2444bd9dcff4c32
Reviewed-on: https://gerrit.chromium.org/gerrit/65569
Reviewed-by: Hsinyu Chao <hychao@chromium.org>
Tested-by: Hsinyu Chao <hychao@chromium.org>
Commit-Queue: Hsinyu Chao <hychao@chromium.org>