CRAS: Move mute/softvol to after loopback.
Moving the zeroing of the mix buffer and the application of the
system/node software volume to after the streams are mixed
accomplishes two things:  It allows the loopback stream to play even
if the system is muted, and ensures that the dsp sees full scale
samples for processing.

BUG=chromium:266547
TEST=capture loopback while the system is muted and make sure there is
audio. Use software volume to adjust HDMI output.

Change-Id: I9aa52bfae52b43bc64524947b47b278cacf06b1d
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/65001