CRAS: Add check_volume API.
Allow a client to request that the volume of a node be checked.  This
will be used to poll USB devices who can change their own volume such
as the Jabra.  The check_volume call will pull the volume from the
device, re-set it (this is a hack specifically for the Jabra), and
updates the node volume to the closest increment to the new level.

BUG=chrome-os-partner:19465
TEST=manual, change volume and use test client to check volume.

Change-Id: I518b667611f6312f95e4fb551211366f868d4b5c
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/66665