CRAS: alsa_helper - Try mmap_begin again on resume.
When mmap_begin fails because the system was suspended, try calling it
again.  Before the error was noted and recovery was run, but the
mmap_begin wasn't called again.  This resulted in a bogus
(uninitialized) pointer being returned since my_areas wasn't
necessarily set before mmap_begin returned an error.

BUG=chromium:290464
TEST=suspend resume while playing youtube.

Change-Id: Ia10344b6b0b3382b749686ac2ea888d7cbd008aa
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/169386
Reviewed-by: Hsinyu Chao <hychao@chromium.org>