peach_pit: Mute Speakers when playing to HDMI.
Config change that mutes the speakers when playing to the HDMI output.
This is needed since the speaker control isn't standard "Speaker" and
instead is "Speaker Left" and "Speaker Right".  There isn't currently a
way to specify a UCM config for a default node like speaker, that will
come in a later change.  This is the simplest way that doesn't touch
common paths close to 30 going stable.

BUG=chromium:22258
TEST=plug/unplug HDMI monitor and toggle with the system tray.

Change-Id: I29f1935573eadcb2895c04fc84082f38f64f0423
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/169661
Reviewed-by: Chih-Chung Chang <chihchung@chromium.org>