CRAS: audio_thread - mute buffer when all streams under run.
If no playback streams provide any data, mute the output instead of
replaying data already in the circular buffer.  Something has changed
that is causing streams to be left running but with their requests for
data unanswered.  This handles that case, by muting instead or repeating
a short buffer of audio.

BUG=chromium:274571
TEST=updated unit tests, play/record audio on bolt.

Change-Id: I547d722c68b3e291fbcf235fbcc4cfd643d78092
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/171659
Reviewed-by: Hsinyu Chao <hychao@chromium.org>