CRAS: loopback - Rework system loopback path
Change the loopback path to keep a buffer instead of sending output data
directly to the loopback client.  This allows for playback streams with
higher latency than the loopback stream.  Currently YouTube has a higher
latency than is used for loopback.  This method adds 4096 samples of
latency to the loopback path.  This could be optimized by making the
buffer size dynamic depending on the latency of all the currently
playing output streams, but that optimization hasn't proven necessary
yet.

BUG=chromium:304346
TEST=record system audio with loopback.

Change-Id: I32a119c1be175333fe788dbd0875cbe45d61387e
Reviewed-on: https://chromium-review.googlesource.com/172302
Reviewed-by: Dylan Reid <dgreid@chromium.org>
Commit-Queue: Dylan Reid <dgreid@chromium.org>
Tested-by: Dylan Reid <dgreid@chromium.org>