CRAS: consume only data in the current buffer for client reading.
The crash happens when the read pointer moves to a wrong place and tries
to read over the end of the buffer:

cras_audio_shm_area: [header][...buf0...][...buf1...]
                                                 ^
                                                 |
                                                 +-- ptr

This could happen if the client reads buf0 too slowly, and
cras_shm_buffer_read() found read_offset > write_offset because the
write_offset is being updated with a value in the middle of buffer,
and it thinks it should consume part of buf1.

The fix is to only consume data in one buffer during client capture
callback, because the client can only read from one buffer at a time.

BUG=chromium:322788
TEST=start two "cras_test_client --capture_file" with different values
of --callback_threshold and verify the client does not access memory
over the end the buffer.

Change-Id: I244f489171f6a53be1ba2afb5ff18d4f8db45203
Reviewed-on: https://chromium-review.googlesource.com/178406
Reviewed-by: Dylan Reid <dgreid@chromium.org>
Tested-by: Chih-Chung Chang <chihchung@chromium.org>
Commit-Queue: Chih-Chung Chang <chihchung@chromium.org>