CRAS: empty_iodev - Simulate capture/playback rate.
Instead of always returning the same buffer levels, simulate the
behavior of a real iodev.  This will fix issues where audio and video
playback are broken when the empty iodev is selected.  Before this audio
would go as fast as possible, confusing the A/V sync and causing video
not to work.

BUG=chrome-os-partner:24118
TEST=aplay/arecord with empty device and make sure duration is correct.

Change-Id: Iabfb00a9562539f4f9373fa2d30d7815e2ed1ed6
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/178721