CRAS: fmt_conv - Convert mono & stereo to 5.1 surroud
The output channel count and layout is fixed when a
6-channel audio is playing, which makes new streams in
mono or stereo unable to be added.
Add the logic to convert mono and stereo to 5.1 surround
output, so that a 6-channel wav file and youtube can be
played at the same time.

BUG=chromium:214597
TEST=Play a 6-channel wav file, and open YouTube at the
same time.

Change-Id: I3c06f242928eae8e1ca1c647a337e7986df44d15
Reviewed-on: https://chromium-review.googlesource.com/179297
Reviewed-by: Dylan Reid <dgreid@chromium.org>
Commit-Queue: Hsinyu Chao <hychao@chromium.org>
Tested-by: Hsinyu Chao <hychao@chromium.org>