cras_rstream, cras_test_client - Change to time based buffer size requirement.
The original threshold is a hard limit: 41 frames, which is 5ms on 8K HZ audio,
and only 0.2 ms on 192K audio. Use it as a globle limit is not reasonable.

Besides, when user play a second stream, it may get converted to match the
current playback. So, setting 480 frames for 192K audio will be changed
to 40 frames after converting to a 16K audio (480 * 16 / 192 = 40).

The playback would fail just because of the frame rate that is being used.

We change it to time based requirement (1 millisecond) instead of a hard limit.

BUG=chromium:307378
TEST=Run the audiovideo_CRASFormatConversion test on link

Change-Id: Ie2ff4e4d11d152a5692924ae95b0b30e46ce4a3c
Reviewed-on: https://chromium-review.googlesource.com/180390
Reviewed-by: Dylan Reid <dgreid@chromium.org>
Commit-Queue: Owen Lin <owenlin@chromium.org>
Tested-by: Owen Lin <owenlin@chromium.org>