CRAS: cras_rclient - Force an even buffer size.
Buffer size is halved to get callback threshold, make sure it is even so
we don't end up off by one for callback threshold.  The callback
threshold is all the client cares about, and it is still allowed to be
odd.  So far odd sizes only happen when sample rate converting.

BUG=chromium:327817
TEST=go to google play and check buffer sizes (it makes both a 48k and
a 44.1 stream).

Change-Id: I6af50bd1d8445e25bdc6ec8904e9f8dc74808692
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/180674
Reviewed-by: Hsinyu Chao <hychao@chromium.org>
Reviewed-by: Chih-Chung Chang <chihchung@chromium.org>