CRAS: fmt_conv - only check layout if nchans > 2
When checking if format conversion is needed, only check the channel
layout if the number of channels is greater than 2.  When creating a
format converter the number of channels is already checked, and it skips
checking layouts if <= 2.  This could lead to a format converter being
create with no conversions, and a segmentation fault when indexing to
the last converter (index -1).

BUG=chromium:327512
TEST=play 44.1 youtube with 44.1 from test_client, no crash.

Change-Id: I98bcd8b63bdedf88e6231581c6f798e3129c1ddc
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/181551
Reviewed-by: Hsinyu Chao <hychao@chromium.org>