CRAS: alsa_plugin - Fix PCM delay before the first callback is received.
snd_pcm_cras_delay can be called before pcm_cras_process_cb is called
for the first time. So pcm_cras->playback_sample_time is still 0, which
causes the function to use an absurdly long delay. Fix this by checking
that playback/capture_sample_time and hw_ptr are all 0, and setting the
delay to 0 in that case.

BUG=chromium:331637
TEST=aplay/arecord audio.

Change-Id: I80845a7fd88b3cb70cae8f359dda84c7af4e36bd
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/181712
Reviewed-by: David Schneider <dnschneid@chromium.org>
Reviewed-by: Hsinyu Chao <hychao@chromium.org>