CRAS: Add a property disable_emphasis to disable emphasis/deemphasis
Emphasis/Deemphasis is going to be removed in Chrome. In order to
align the behavior of dsp and chrome, we need to enable
emphasis/deemphasis if user is using an old chrome version to
calibrate config files, and disable it if user is using a new chrome
version to calibrate config files.
This patch adds a property disable_emphasis to drc to disable
emphasis/deemphasis.

CQ-DEPEND=CL:186340
BUG=chromium:340870
TEST=1. Prepare a audio raw file sweeping from 10Hz to 22000Hz.
     2. Modify parameters in drc_test so three drc filters will clip
       audio that is larger than 0dB. Specifically, modify RATIO to
       999, KNEE and THRESHOLD to 0, POST_GAIN to 0. Also, add
       drc->emphasis_disabled = 1;
     3. run drc_test and check the waveform of output. The magnitude
       at high frequency remains roughly the same if
       drc->emphasis_disabled = 1. The magnitude at high frequency
       will be lowered if drc->emphasis_disabled = 0.

Change-Id: I98ff15a88fd198d0239fe1fa3f324eb0875642ba
Reviewed-on: https://chromium-review.googlesource.com/186073
Reviewed-by: Chih-Chung Chang <chihchung@chromium.org>
Commit-Queue: Cheng-Yi Chiang <cychiang@chromium.org>
Tested-by: Cheng-Yi Chiang <cychiang@chromium.org>