scripts: Fix audio-tuning frontend createMediaElementSource error
When audio file is played in a loop, we don't need to reset source node
in every loop. Now we just set the source node when there is a new file
url being set.

BUG=chrome-os-partner:26686
TEST=run audio-tuning frontend, check there is no error message
complaining audioContext.createMediaElementSource gets an invalid
argument. Also check that processing effect remains the same
when playing a file in a loop.

Change-Id: Ia75b732690a0d467be0c5096e78bef327780fd18
Reviewed-on: https://chromium-review.googlesource.com/189567
Reviewed-by: Chih-Chung Chang <chihchung@chromium.org>
Commit-Queue: Cheng-Yi Chiang <cychiang@chromium.org>
Tested-by: Cheng-Yi Chiang <cychiang@chromium.org>