CRAS: Track number of active streams of each direction.
Currently we only track and provide number of total active streams
to dbus user. We want to expose a new API to provide number of active
streams of different directions. This CL changes num_active_streams
in cras_server_state struct from an uint32_t to an array.
The setter and getter method are provided in cras_system_state.
Also, the elements in the array is updated when a stream is connected
or removed.

BUG=chromium:336459
TEST=make check

Change-Id: I74328398a216c5fa277ef88935d341d31d18bb97
Reviewed-on: https://chromium-review.googlesource.com/191244
Reviewed-by: Chih-Chung Chang <chihchung@chromium.org>
Commit-Queue: Cheng-Yi Chiang <cychiang@chromium.org>
Tested-by: Cheng-Yi Chiang <cychiang@chromium.org>