CRAS audio_thread - Send longest timeout length to UMA
Add a pipe for audio thread to send message to main thread.
Note that the existing to_main_fds is used for synchronous
message reply sent from audio thread to main thread, and the new
main_msg_fds is for message sent to main thread.  Right now
it is used for sending audio thread metrics to UMA.

If a stream timeout within the 10 seconds limit, keep
track of the longest timeout period length in milliseconds
and log this value when the stream is removed.

BUG=chromium:327817
TEST=Manual modify cras_test_client to create a bad
stream, verify histogram in about:histograms

Change-Id: I34fa220db6cbefdba3629070af8a5e3f39e741b9
Reviewed-on: https://chromium-review.googlesource.com/192110
Reviewed-by: Dylan Reid <dgreid@chromium.org>
Commit-Queue: Hsinyu Chao <hychao@chromium.org>
Tested-by: Hsinyu Chao <hychao@chromium.org>